+ Resolved compile warnings. + Added dependencies to VC80 projects - external libs/includes no longer needed. If you use other than 8th release, modify your projects accordingly.

+++ trunk_4574/src/ascent-world/ConsoleListener.cpp
@@ -187,7 +187,7 @@ void TestConsoleLogin(string& username, string& password, uint32 requestid);
 
 void ConsoleSocket::OnRead()
 {
-	uint32 readlen = GetReadBuffer().GetSize();
+	uint32 readlen = uint32(GetReadBuffer().GetSize());
 	uint32 len;
 	char * p;
 	if( ( readlen + m_pBufferPos ) >= m_pBufferLen )
