+++ trunk_4574/src/ascent-world/Object.cpp
@@ -2445,7 +2445,7 @@ void Object::SpellNonMeleeDamageLog(Unit *pVictim, uint32 spellID, uint32 damage
 		{
 			std::list< DamageSplitTarget >::iterator itr;
 			Unit * splittarget;
-			uint32 splitdamage, tmpsplit;
+			int32 splitdamage, tmpsplit;
 			for( itr = pVictim->m_damageSplitTargets.begin() ; itr != pVictim->m_damageSplitTargets.end() ; itr ++ )
 			{
 				// TODO: Separate damage based on school.
@@ -2458,7 +2458,7 @@ void Object::SpellNonMeleeDamageLog(Unit *pVictim, uint32 spellID, uint32 damage
 						tmpsplit = float2int32( res ); // prevent < 0 damage
 					splitdamage = tmpsplit;
 					res -= (float)tmpsplit;
-					tmpsplit = itr->m_pctDamageSplit * res;
+					tmpsplit = int32(itr->m_pctDamageSplit * res);
 					if( tmpsplit > float2int32( res ) )
 						tmpsplit = float2int32( res );
 					splitdamage += tmpsplit;
